Move key sharing related functions to a new file
